Changed generic

+++ src/main/java/cat/udl/eps/softarch/mypadel/repository/MatchRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface MatchRepository extends PagingAndSortingRepository<Match, Integer> {
+public interface MatchRepository extends PagingAndSortingRepository<Match, Long> {
 }
